requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
flower==1.2.0
humanize==4.3.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.14.0
kombu==5.2.4
packaging==21.3
pkgutil-resolve-name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
